+++ pipeline/pipeline.go
@@ -3,72 +3,87 @@
 package pipeline
 
 import (
+	"errors"
 	"fmt"
+	"strings"
 
 	"github.com/julian7/goshipdone/ctx"
-	// register internal modules
-	_ "github.com/julian7/goshipdone/internal/modules"
-
-	"github.com/julian7/goshipdone/modules"
 	"gopkg.in/yaml.v3"
 )
 
-// BuildPipeline represents the whole build pipeline. Module objects
-// are loaded into BuildPipeline as read from .pipeline.yml.
-type BuildPipeline struct {
-	Setups    *modules.Modules `yaml:",omitempty"`
-	Builds    *modules.Modules `yaml:",omitempty"`
-	Publishes *modules.Modules `yaml:",omitempty"`
+// Pipeline is a generic pipeline, with a registry and stages configured.
+type Pipeline struct {
+	Stages []*Stage
 }
 
-// LoadBuildPipeline creates a new BuildPipeline by reading YAML
-// contents of a byte slice. Then, it makes sure default modules
-// are loaded, providing safe defaults.
-func LoadBuildPipeline(ymlcontent []byte) (*BuildPipeline, error) {
-	pipeline := &BuildPipeline{
-		Setups:    modules.NewModules("setup"),
-		Builds:    modules.NewModules("build"),
-		Publishes: modules.NewModules("publish"),
-	}
+func New(stages []*Stage) *Pipeline {
+	pip := &Pipeline{Stages: make([]*Stage, 0, len(stages))}
+	pip.Stages = append(pip.Stages, stages...)
 
-	pipeline.Publishes.SkipFn = func(context *ctx.Context) bool {
-		return !context.Publish
-	}
-
-	err := yaml.Unmarshal(ymlcontent, pipeline)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, itemType := range []string{"env", "project", "git_tag", "skip_publish"} {
-		_ = pipeline.Setups.Add(itemType, nil, true)
-	}
-
-	return pipeline, nil
+	return pip
 }
 
-// String returns a string representation of the build pipeline
-func (pipeline *BuildPipeline) String() string {
-	return fmt.Sprintf(
-		"{Setups:%v Builds:%v Publishes:%v}",
-		pipeline.Setups,
-		pipeline.Builds,
-		pipeline.Publishes,
-	)
+// UnmarshalYAML parses YAML node to load its modules
+func (pip *Pipeline) UnmarshalYAML(node *yaml.Node) error {
+	if node.Kind != yaml.MappingNode {
+		return errors.New("pipeline definition is not a map")
+	}
+
+	l := len(node.Content)
+	for i := 0; i < l; i += 2 {
+		var stage *Stage
+		stageDefName := node.Content[i].Value
+
+		for _, st := range pip.Stages {
+			if st.Plural == stageDefName {
+				stage = st
+			}
+		}
+
+		if stage == nil {
+			continue
+		}
+
+		if err := node.Content[i+1].Decode(stage); err != nil {
+			return fmt.Errorf("decoding %s stage: %w", stage.Name, err)
+		}
+	}
+	return nil
+}
+
+// LoadDefault loads a module into a stage, if not loaded yet
+func (pip *Pipeline) LoadDefault(kind string) error {
+	items := strings.SplitN(kind, ":", 2)
+	if len(items) != 2 {
+		return fmt.Errorf("invalid module kind: %q", kind)
+	}
+
+	if stg := pip.StageByName(items[0]); stg != nil {
+		if err := stg.Add(items[1], nil, true); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+func (pip *Pipeline) StageByName(name string) *Stage {
+	for _, stage := range pip.Stages {
+		if stage.Name == name {
+			return stage
+		}
+	}
+
+	return nil
 }
 
 // Run executes build pipeline, calling Run on all
 // Modules
-func (pipeline *BuildPipeline) Run() error {
-	steps := []*modules.Modules{
-		pipeline.Setups,
-		pipeline.Builds,
-		pipeline.Publishes,
-	}
+func (pip *Pipeline) Run() error {
 	ctx := ctx.New()
 
-	for _, step := range steps {
-		if err := step.Run(ctx); err != nil {
+	for _, stg := range pip.Stages {
+		if err := stg.Run(ctx); err != nil {
 			return err
 		}
 	}